[add] add BaseI_extension.rs to inst_32
